Knockback added and updated enemy movement
NOTE: I tried to update to a move_and_collide() or move_and_slide() based movement, but am having issues with the CharacterBody2D being a child node to the script-bound main Node2D node. This would cause me to have to do weird things like manually sync the main node to the CharacterBody2D every frame, leading to strange behaviors and the inability to correctly tell if shots are colliding. I have reverted those changes for now and will investigate a better way to do it in the future.